update solana chain code
